+++ BILBO-firmware.X/bilbo/low_level/user_leds.c
@@ -95,6 +95,6 @@ void handle_range_leds_out(uint8_t range){
 #endif
 }
 
-int led_init(){
+uint8_t led_init(){
     LED_GREE_Set();
 }
